% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_columns.R
\name{align_columns}
\alias{align_columns}
\title{Align columns of multiple shapefiles}
\usage{
align_columns(shapefiles)
}
\arguments{
\item{shapefiles}{list of shapefiles}
}
\value{
list of aligned shapefiles
}
\description{
This function ensures that all shapefiles have the same number of columns
filling with NA's where needed.
all_columns represents unique columns present in all shapefiles and
missing_columns find missing columns in a shapefile and fills with NA
}
