% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diary.R, R/create_diary_from_ogt.R
\name{create_diary}
\alias{create_diary}
\title{Create fishing diary from two OGT devices}
\usage{
create_diary(df1, df2, threshold)

create_diary(df1, df2, threshold)
}
\arguments{
\item{df1}{Data frame with the gps info from the first OGT}

\item{df2}{Data frame with the gps info from the second OGT}

\item{threshold}{Time difference in seconds between two measures that we
consider before marking a point, time less that this amount is considered
spurious or false and is not taken into account}
}
\value{
A composed data frame with the

A composed data frame with the
}
\description{
This function combines the two files coming from two OGT devices into a single
one where operations are marked. Understanding how OGTs work is essential to
understand what this function is doing.
From these two files start and end of setting and hauling events are
identified.
Before using this function, data from OGTs has to be processed with
process_white_gps_data() function from this library.

This function combines the two files coming from two OGT devices into a single
one where operations are marked. Understanding how OGTs work is essential to
understand what this function is doing.
From these two files start and end of setting and hauling events are
identified.
Before using this function, data from OGTs has to be processed with
process_gps_data() function from this library.
}
\examples{
create_diary(ogt1, ogt2, 600)
create_diary(ogt1, ogt2, 600)
}
