% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_layer_with_offset.R
\name{crop_layer_with_offset}
\alias{crop_layer_with_offset}
\title{Crop sf object based on bounding box of another sf object}
\usage{
crop_layer_with_offset(layer_to_crop, bbox_layer, offset = c(0, 0, 0, 0))
}
\arguments{
\item{layer_to_crop}{Sf object that is going to be cropped.}

\item{bbox_layer}{Layer from which the bounding box is obtained.}

\item{offset}{Desired offset to apply to the bounding box.}
}
\value{
A cropped sf object.
}
\description{
This function crops an sf object based on the bounding box calculated from another sf objetc
and a desired offset, that may be applied to the four sides of the bounding box or may be specified
for each side.
When specifying your offset take into account if sf objects are projected or not.
}
\examples{
# Crop specifying different offsets for each side, order is xmin, xmax, ymin, ymax.
cropped_layer <- crop_layer_with_offset(sf_object_to_crop, sf_object_from_which_bbox_is_obtained, offset = c(0.01, 0.01, 0.2, 0.01))
# Crop specifying global offset.
cropped_layer <- crop_layer_with_offset(sf_object_to_crop, sf_object_from_which_bbox_is_obtained, offset = 10)
}
