% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_match_oper.R
\name{gps_match_oper}
\alias{gps_match_oper}
\title{Assign operations using gps and observer or OGT operations diary.}
\usage{
gps_match_oper(gps_file, log_file, timezone = "UTC", log_source = "obs")
}
\arguments{
\item{gps_file}{File where the gps data is.}

\item{log_file}{File with the diary info.}

\item{timezone}{Defaults to UTC.}

\item{log_source}{Defines where diary data comes from. Options are "obs"
for observer, "ogt" for OGTs and "log" for captains log. Defaults to "obs".}
}
\value{
returns a gps file but with fishing operations marked in a new
column named SI_FOPER and values ST (Steaming), SE (Setting), HA (Hauling)
and WT (Waiting), SI_FSTATUS marked as TRUE or FALSE and the corresponding
SU_ISOB, SI_OGT or SI_LOG filled where TRUE.
}
\description{
Fishing diary can be generated by an observer,by create_diary_from_ogt
function or by captains log.
For diaries to work, some columns need to be present:
Fecha
fec_cal
ini_largada
fin_largada
fec_vir
ini_virada
fin_virada
}
\details{
These columns will be automatically generated if create_diary_from_ogt is
used, so make sure that these columns are present in the file if the log file
is created by an observer. On ly dates and times are needed, no positions
have to be included on this file.
}
