% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shapefiles.R
\name{load_shapefiles}
\alias{load_shapefiles}
\title{Load all .shp files in a directory}
\usage{
load_shapefiles(directory, combine = FALSE, target_crs = NULL)
}
\arguments{
\item{directory}{directory where shapefiles are stored.}

\item{combine}{sets if shp files has to be combined into a single shapefile. Defaults to FALSE}

\item{target_crs}{converts shapefiles to a common crs. If no crs is given, all will be
transformed to the crs of the fisrt shapefile processed. Otherwise a crs can be given.}
}
\value{
A list containing all the shapefiles or a single combined shapefile.
}
\description{
This function loads all the .shp files inside a directory using a safe
function that will continue loading files if errors or warnings are thrown.
}
\examples{

Define directory where.shp files are located
shapefile_directory <- "/home/user/shapefiles_dir/"
Load and combine shapefiles.
combined_shapefile <- load_shapefiles(shapefile_directory, combine = TRUE, target_crs ="epsg:25830")
Load shapefiles without combining.
shapefiles_list <- load_shapefiles(shapefile_directory, combine = FALSE)
}
