% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_harbour.R
\name{mark_harbour}
\alias{mark_harbour}
\title{Mark vessel point inside or outside a harbour}
\usage{
mark_harbour(
  vessel_track_sf,
  harbour_dataframe,
  lon_col = "LON",
  lat_col = "LAT",
  range_col = "RANGO"
)
}
\arguments{
\item{vessel_track_sf}{An sf object with vessel tracks.}

\item{harbour_dataframe}{A dataframe with the location (longitude, latitude) of the harbour and a range column (in km) inside which we consider the vessel is inside the harbour.}

\item{lon_col}{Name of the longitude column in the harbour_dataframe Default is "LON".}

\item{lat_col}{Name of the latitude column in the harbour_dataframe Default is "LAT".}

\item{range_col}{Name of the range column in the harbour_dataframe Default is "RANGO".}
}
\value{
An sf object with the SI_HARB (TRUE or FALSE values) column added.
}
\description{
This function creates the column SI_HARB that indicates if the point is inside (TRUE) or outside (FALSE) a harbour.
}
