% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Reader wrapper to read csv or RData files}
\usage{
read_file(file_path)
}
\description{
This function allows to use both csv and RData files where needed. It is
intended to use inside another function so no need to do direct calls
}
