% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saferead_shapefile.R
\name{saferead_shapefile}
\alias{saferead_shapefile}
\title{Safely read a shapefile}
\usage{
saferead_shapefile(file)
}
\arguments{
\item{file}{Path to the shapefile.}
}
\value{
A list with succesfully loades shapefile.
}
\description{
This function loads a shapefile and manages warnings and errors.
}
