% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_match_oper.R
\name{gps_match_oper}
\alias{gps_match_oper}
\title{Assign operations using gps and observer or OGT operations diary.}
\usage{
gps_match_oper(gps_file, log_file, timezone = "UTC", use_aggregated = FALSE)
}
\arguments{
\item{gps_file}{File where the gps data is.}

\item{log_file}{File with the diary info.}

\item{timezone}{Defaults to UTC.}

\item{use_aggregated}{Defines if diary data has to be aggregated by gear,
This has to be set to TRUE if vessel has used more than one gear in a day,
but defaults to FALSE}
}
\value{
returns a gps file but with fishing operations marked in a new
column named SI_FOPER and values ST (Steaming), SE (Setting), HA (Hauling)
and WT (Waiting)
}
\description{
Fishing diary can be generated by an observer or by create_diary_from_ogt
function.
For diaries to work, some columns need to be present:
Fecha
fec_cal
ini_largada
fin_largada
fec_vir
ini_virada
fin_virada
}
\details{
These columns will be automatically generated if create_diary_from_ogt is
used, so make sure that these columns are present in the file if the log file
is created by an observer. On ly dates and times are needed, no positions
have to be included on this file.

One advantage is that function can process all .csv files under a directory or can do the process for single files.
Also a gear for the vessel may be specified if it is known.
}
